+++ lib/terradactyl/terraform/commands/init.rb
@@ -101,6 +101,12 @@ module Terradactyl
       end
     end
 
+    module Rev1_latest
+      module Init
+        include Terradactyl::Terraform::Rev015::Init
+      end
+    end
+
     module Commands
       class Init < Base
       end