+++ lib/terradactyl/terraform/planfile.rb
@@ -148,6 +148,11 @@ module Terradactyl
       end
     end
 
+    module Rev1_latest
+      class PlanFileParser < Rev012::PlanFileParser
+      end
+    end
+
     module Rev011
       class PlanFileParser < Rev012::PlanFileParser
         def checksum